Significant changes and bootstrapping...
We now have a much better hook into the read
table, and a way to easily define data literals
using reader:push-reader!  .  This was actually
tricky, since we needed to expand the notion of
' quotation to include canonical CL forms, but
also enable the quoted form to be read as a
clojure structure as well.  This scheme will
buy some mileage for now.  Still need to
hack the reader to coerce ~ to , and ~@ to
,@ for clojure forms (or at least make them
compatible, possibly with a pre-processing
reader macro....)

Currently, only pvector is implemented, and
it works with quotation now.

Protocols exist, as do multi-arity fns, although
I've only been testing single-arity...

Looking to move toward a CLOS-based function
object representation, closer to clojure's
notion of class-based closures and generalization
of metadata.  CLJS may be more helpful here due
to it's nearer-ness to CLOS; dunno yet.  There's
a working example of a function object, or fob,
at the bottom of bootstrap.lisp

I'm shifting away from writing a custom eval, and
more toward just macroizing the hell of out everything.
Have little idea about interop.  Going this way should
allow us to splice in CL forms if we want to though
(particularly if we can toggle cl/clj reading with a
dynamic var or even a dispatch char).

I need to get the Seq/ISeq protocols in soon, and
build on them.  Most of the core libs use seqs
pervasively.  This will require redefining apply to
accept things other-than-lists as args via the seq
abstraction.

Overall, looking pretty good!