Big strides in variadic args.  Also keyword access
Defining generic functions, validating args,
and storing meta about the arity and args.

Just need to implement the method definitions,
which confer with the arities to validate
the implementation and map bodies to arity
for dispatch.

Also figured out a potential way to do
keyword access within the constraints of the
system (promising, worked at the repl).