Working on getting evaluation semantics for vecs
The reader implementation of vectors returns
unevaluated vector literals.  That is expected,
except we need to evaluate the vector if it's
in a non-read context, like the right hand side
of a binding form, or the result of any body
of the special forms.

Current strategy is to hack in clj-eval,
which has an evaluation rule for vectors
(and later maps, sets, any other literals)
so that we can build our primitive forms
like let on top of it.  We can provide
evaluation for the rhs of bindings, and
for the body.  clj-eval wraps cl:eval,
so we should not be playing with fire too much.
This is the least invasive strategy at the
moment.