Working on some degenerate reader cases
Working on adding metadata examples,
using new cowmaps, ran into the problem
that

{:tag 'blah}

tries to print as
{:tag QUOTE BLAH}

This also led to realization that

(def x 2)

[:a x]

yields
[:a x]

instead of
[:a 2]

Evaluation semantics are getting a little
janked when we read, likely due to syntax quote.

For the map, flatten is causing an ostensibly
good read to go sour, and decoupled the quote
into a triple (should only be a pair).

So we need to police our symbols at read time.
I think this is circling back to the original
problem that showed up.