Added hacked evaluation and literals
Patched some functionalit into eval.lisp
that allows user to control evaluation
semantics in sbcl.  User can extend
a multimethod for custom types
custom-eval, which will serve as
a final hook in the non-list eval
dispatch.  Default is identity (standard
sbcl behavior).  User can enable or
disable at will.  Hack is somewhat minor,
but requires unlocking the sb-imp
package to do so.  Should be of
limited scope and use for bootstrapping.

There's some disconnect with the
reader and quasiquoting that I'm
trying to work out.