Looks like quasiquoting is working
I hacked together an internal implementation
of a quasi-quoted read for the reader macros.
This currently relies on sbcl internals; no
idea if it's portable but I don't really care
for bootstrapping.  It looks like we
get quasiquoting semantics with our custom
data literals now.

Might be a good idea to roll custom eval,
quoting, read, up into a literal-data
macro that spits all this stuff out for us
since it's taking more-or-less the same
form each time.  The downside is we're
punting and building up arglists to feed
to the data constructor each time (eager).

In a production system this would be undesirable
and we'd probably want to use sequences (maybe not).
For limited purposes, I think it will be fine.