Fixed package-specific interning by defining scase
Due to the funkiness of package symbol interning
causes eql to yield false in uses of #'case for
packages where the function was not defined
due to qualified symbols, I wrote a new
predicate to provide portable symbol equality
closer to clojure's semantics.

common-utils:symbol? determines if
an arg is definitely a symbol, excluding
keywords.

common-utils:symbol= defines a predicate
that yields t if both args pass symbol? and
the symbol-names are string=.

common-utils:seql is a replacement for common-lisp:eql,
where if the args are symbol?, we use symbol=, otherwise
we test using eql.

common-utils:custom-case allows callers to define case
macros with custom tests.

common-utils:scase defines a drop-in replacement for
common-lisp:case that uses seql instead of eql for
the test condition, meaning that we can know treat
symbolic tests as generic, unqualified symbols.  This
enables functions like detect-recur and other code walking
functions e.g. for with-recur, to work in other namespaces
without the interning garbage requirements.  This should
be a generally acceptable practice, unless I'm missing something
critical.  The current implementation largely wraps sbcl internals,
although one could be written more portably.  This is a non-concern
for bootstrapping though.