Temporarily disabled literal eval rules
After the last round of custom eval
rules, we ran into a problem.

To comport to CL's expectation of
sexpr forms for lexical analysis
(data literals - not so good),
we do a little pre-eval
pass over any forms in let bindings
or similar binding forms, and
apply custom eval rules as
defined by a couple of methods.

This allows us to coerce our
literal structures (like vectors)
into "safe" sexprs, which
SBCL will happily compile.

The downside is that this technique
will also transform the literal
for macros (if inside a let),
so that code like Protocols, which
expects an actual vector literal
for the args, and any other
macrology that uses literal vectors,
has a chance to end up as a
(persistent-vector ...) sexpr
instead of the actual literal.

We ended up violating some
assumptions in clclojure.protocols,
and causing errors related to
applying functions to sexprs instead
of vector literals.

For now, I've disabled the transform
for vectors to enable a reversion
to a functioning form.

The goal of course is to unify
custom eval semantics, without
forcing the transformation for things like
macros.

Current avenues for addressing this problem
are to define a (literal ...) form that
evals to identity, but serves as a simple
tag that we can walk and detect with macros,
replacing with vector literals at macroexpansion
time.

Another option could be to to preclude the eval-time
code transformation inside the bodies of macro expressions,
deferring that until the macro goes to eval.

Yet another (deeper) would be to hack the compiler
and teach it about custom data literals.