Developed inline variadic named functions
Since we no longer use labels for the
named functions and invoke them,
we now just have a single labels to
allow for self recursion, and
have the arg-dispatch invoke
a tailored with-recur form
relevant to the arg arity.

Seems to work so far,
should be compatible with
lazy-seq as well.