Big overhaul, migration to new impl
We have most of the original boostrap
from the earlier implementation ported
over, to the new CL native runtime.

We also have interop with native
clojure keywords,vars,symbols, and
early namespaces.

Next step will be to leverage
the runtime to try to get reader
support up, then analyze/emit/eval.

Lacking multimethods, but that should
not be terribly hard to pull if there's
not already a library for it.