Continuing with port, fleshing out str and types
We have -to-string under IString now, so
str is generic in base.

Forgot that I already with-slots inside of
clojure-deftype bodies, so we have slot access
compatible with setf.
added a macro set! that just wraps setf for
parity.

Having case of dueling constructors at compile
time, where the compiler is sad with records
since clojure-deftype emits a ->type ctor,
which is equivalent to type. from clj.
defrecord also does its ctor as ->, so
this redefines well at runtime, but compile
time throws false positive for different
args (e.g. due to meta and hash fields).

so we want to disambiguate.  we can just
emit type. as a defun instead of the ->ctor
and let defrecord have it.

We should provide symbol-macrolet to coerce
type. into (make-instance ....) inside
function bodies though (for parity with
clj).