Found workaround for dynamic binding collisions
Tried using deflex but it didn't answer the
mail.

Better solution is to leverage setq for top
level binds (some say this can go screwy,
we'll see).

So now def will dispatch on the name, and
if it's earmuffed, we use defparamter otherwise
(default) we use setq.

with-slots is happy again, and our deftype
definitions are cosistent with clojure's
semantics.  The only question remaining is
how to reference the outer lexical environment
if we "really" want to.  I don't think we
can get it outside of an indirect resolution
mechanism.  Unsure if we can in clojure even...