Edging even closer to destructuring, more unified
Defined unified-lambda to allow us to
have lambda forms that can either be
destructured or not, where their arguments
are unified in a lexical lisp1 akin to
unified-let*.  This was a minor oversight
from before, but it allows us to pass in
args and funcall them directly (from the
caller's perspective).

We support normal lambda lists and
destructuring lists, trying to detect
either.  If a destructuring list is
used, we leverage metabang-bind:bind
and replace-funcalls (over an analyzed
var extraction of the params) to get
our lisp1 context.

Also rewrote the definition of fn
to be able to deal with destructured
params.  Where we had a simple nested
list detection scheme before, we have
to do a bit more work to determine
the input arg-body specs due to
nested list possibilities from
destructuring.  This will be WAY
easier when we have vector literals
to communicate bindings, but for now
we figured out some invariants to help
detect destructured binding specs.

The next step is to get named-fn,
named-fn* working with destructuring.

Notably, both leverage labels to establish
named recursion sites (in addition to
the extant with-recur recur semantics).

We return the #'name function defined
by the label.  An interesting consequence
is that we can't pass this around as
a function object like we can lambdas.

It might behoove us to wrap the label fn
in a lambda we return.  That lambda could
leverage unified-lambda to handle
destructuring as well (although we have
a little snag with the lexical package
depending on common-utils, might be surmountable
though).

Lots of progress though.