Lots of progress on function binds
We have a pretty decent destructuring
in function binds now.  Close to
parity with the seq/vector based
forms in clojure, enough to port
them straightfowardly.

We can probably implement keys as well.
Currently lacking :as, but that looks
pretty trivial to implement (we already
provide temp vars for the parent binding,
so it's just a matter of scraping these
out.