Working again, all ambiguous functions parsing.
Had to revise and make consistent the
format we're communicating function specs,
particularly function bodies, wherein we
had special cases depending on atoms or not.

We now consistently process lists from
the parse results for a body, where we
assume an implicit progn is available.

Then we dispatch as needed down the line.

Found a lot of littl corner cases with
"" and nil/() where parsing assumptions
were failing. We appear to be robust enough
to compile now though!