Back to working state, integrating cons.
Unintentionally broke cons by redefining it
after using it.  seq functions were borked,
and printing of plists was dependent on them,
so borked.

we were using a cons defintion for inverted
conj which worked fine until now.  we now
have an intentional construct for creating
persistent lists.

So we either need to replace all the prior
cons calls with this and get it to work,
or we need something else.

original cos was polymorphic on -conj, so
worked with CL objects too.

I think we should do something like that
with cons, but have it work with the persistent
list route cleanly.

problem arises if we want to cons to nil.
do we make a cl cons or a clojure persistent
list?  How far do we go, or which default do we
choose.

Original instinct: we default to cons nil yielding
a persistent pathway, with seqs.