Patched sequence fixes in, unified legacy stuff
Had a minor conundrum to port all the legacy
cl-native implementations from sequences.lisp
that I'd built on top of.  In the final
implementation, all we need is defprotocol
and deftype to implement the cljs core which
includes sequence stuff.  For mild expediency
and pedagogical reasons, I opted to currently
retain implemnations in sequences.lisp, which
also includes early terminating reductions
and internal-reduce.  we need to unify this
stuff eventually though, but for now, we
provid default method implmentations for
the genric object type t, which in turn
dispatches to clojure's relative protocol
impls.  So we have backwards and forwards compat.

It's mildly ugly right now, but everything
is seqable and reducible so far (includinng
built in CL types).