Refactored clj-objects
Shifted toward some superclasses
for clojure primitive colls to implement
so we get meta/hasheq baked in.

might be simpler to decouple all the slots but
meh, this works and is leveraging CLOS facilities.

Also patched around a small bug in the seq
impl where we didn't have internal-reduce
defined, it will patch into -reduce for
all types now as a catch-all.

our basic operations on seqs are working again.